Pipeline + diff process
