+++ source/app.ts
@@ -7,9 +7,6 @@ import fs from 'fs';
 import './config';
 import router from './api/router';
 import { getLogger } from './helpers/logger';
-import { prepareExportData } from './helpers/promExporter';
-import redis from './db/redis';
-import { cleanJob } from './db/redis-di';
 
 const app = express();
 
@@ -38,30 +35,11 @@ const logger = getLogger('app');
 createServer(app).listen(PORT, IP, () => {
     logger.info(`Started tracker at ${IP}:${PORT}`);
 }).on('connection', (socket) => {
-    const timeout = 10 * 1000; // 10s timeout. A bit harsh but normall it takes < 10ms for proper announce.
+    const timeout = 2 * 1000; // 2s timeout. A bit harsh but normall it takes < 10ms for proper announce.
     socket.setTimeout(timeout);
     socket.once('timeout', () => {
         socket.end();
     });
-})
-
-setInterval(() => {
-    cleanJob(redis);
-}, 1000 * 60 * 5); // Every 5 minutes
-
-const PROM_IP = process.env.PROM_IP || '127.0.0.1';
-const PROM_PORT = parseInt(process.env.PROM_PORT || '9999');
-const promApp = express();
-
-promApp.get('/metrics', async (req, res) => {
-    const responseText = await prepareExportData();
-    res.set('Connection', 'close');
-    res.end(responseText);
-    res.socket?.end();
-})
-
-createServer(promApp).listen(PROM_PORT, PROM_IP, () => {
-    logger.info(`Started prometheus metrics server at ${PROM_IP}:${PROM_PORT}`);
 })
 
 process.on('SIGINT', () => {
